requirements.txt
pyglet
pycryptodome